RAG/intro-to-vector-dbs/pyproject.toml
[project]
name = "intro-to-vector-dbs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-google-genai>=2.1.12",
    "langchain-pinecone>=0.2.12",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.1",
]
